+++ frontend/src/App.css
@@ -23,6 +23,11 @@ aside a.active span{
   color:white;
   background-color: #ef4444;
 }
+/* category Nav */
+ a.active .cat{
+  color:white;
+  background-color: #ef4444;
+}
 
 section{
  grid-column-start: 3;
