--- frontend/src/components/book/test.jsx
@@ -1,373 +1,0 @@
-{
-    "message": "Book Fetched Successfully",
-    "Book": {
-        "kind": "books#volumes",
-        "totalItems": 6,
-        "items": [
-            {
-                "kind": "books#volume",
-                "id": "PNJZAAAAcAAJ",
-                "etag": "/XxKmaDDQ/E",
-                "selfLink": "https://www.googleapis.com/books/v1/volumes/PNJZAAAAcAAJ",
-                "volumeInfo": {
-                    "title": "The School for Daughters. A Drama, Etc",
-                    "authors": [
-                        "Dennis Lawler"
-                    ],
-                    "publishedDate": "1808",
-                    "industryIdentifiers": [
-                        {
-                            "type": "OTHER",
-                            "identifier": "BL:A0018116962"
-                        }
-                    ],
-                    "readingModes": {
-                        "text": false,
-                        "image": true
-                    },
-                    "pageCount": 92,
-                    "printType": "BOOK",
-                    "maturityRating": "NOT_MATURE",
-                    "allowAnonLogging": false,
-                    "contentVersion": "0.4.4.0.full.1",
-                    "panelizationSummary": {
-                        "containsEpubBubbles": false,
-                        "containsImageBubbles": false
-                    },
-                    "imageLinks": {
-                        "smallThumbnail": "http://books.google.com/books/content?id=PNJZAAAAcAAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
-                        "thumbnail": "http://books.google.com/books/content?id=PNJZAAAAcAAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api"
-                    },
-                    "language": "en",
-                    "previewLink": "http://books.google.com/books?id=PNJZAAAAcAAJ&printsec=frontcover&dq=/pnJZAAAAcAAJ&hl=&cd=1&source=gbs_api",
-                    "infoLink": "https://play.google.com/store/books/details?id=PNJZAAAAcAAJ&source=gbs_api",
-                    "canonicalVolumeLink": "https://play.google.com/store/books/details?id=PNJZAAAAcAAJ"
-                },
-                "saleInfo": {
-                    "country": "US",
-                    "saleability": "FREE",
-                    "isEbook": true,
-                    "buyLink": "https://play.google.com/store/books/details?id=PNJZAAAAcAAJ&rdid=book-PNJZAAAAcAAJ&rdot=1&source=gbs_api"
-                },
-                "accessInfo": {
-                    "country": "US",
-                    "viewability": "ALL_PAGES",
-                    "embeddable": true,
-                    "publicDomain": true,
-                    "textToSpeechPermission": "ALLOWED",
-                    "epub": {
-                        "isAvailable": false,
-                        "downloadLink": "http://books.google.com/books/download/The_School_for_Daughters_A_Drama_Etc.epub?id=PNJZAAAAcAAJ&hl=&output=epub&source=gbs_api"
-                    },
-                    "pdf": {
-                        "isAvailable": false
-                    },
-                    "webReaderLink": "http://play.google.com/books/reader?id=PNJZAAAAcAAJ&hl=&source=gbs_api",
-                    "accessViewStatus": "FULL_PUBLIC_DOMAIN",
-                    "quoteSharingAllowed": false
-                }
-            },
-            {
-                "kind": "books#volume",
-                "id": "PNjzAAAACAAJ",
-                "etag": "iFoJqGL/d08",
-                "selfLink": "https://www.googleapis.com/books/v1/volumes/PNjzAAAACAAJ",
-                "volumeInfo": {
-                    "title": "Retro Book",
-                    "authors": [
-                        "Advantage Publishing NSW Pty, Limited"
-                    ],
-                    "publisher": "Highbury Australia",
-                    "publishedDate": "2004-01-01",
-                    "industryIdentifiers": [
-                        {
-                            "type": "ISBN_10",
-                            "identifier": "0975686267"
-                        },
-                        {
-                            "type": "ISBN_13",
-                            "identifier": "9780975686263"
-                        }
-                    ],
-                    "readingModes": {
-                        "text": false,
-                        "image": false
-                    },
-                    "printType": "BOOK",
-                    "maturityRating": "NOT_MATURE",
-                    "allowAnonLogging": false,
-                    "contentVersion": "preview-1.0.0",
-                    "language": "en",
-                    "previewLink": "http://books.google.com/books?id=PNjzAAAACAAJ&dq=/pnJZAAAAcAAJ&hl=&cd=2&source=gbs_api",
-                    "infoLink": "http://books.google.com/books?id=PNjzAAAACAAJ&dq=/pnJZAAAAcAAJ&hl=&source=gbs_api",
-                    "canonicalVolumeLink": "https://books.google.com/books/about/Retro_Book.html?hl=&id=PNjzAAAACAAJ"
-                },
-                "saleInfo": {
-                    "country": "US",
-                    "saleability": "NOT_FOR_SALE",
-                    "isEbook": false
-                },
-                "accessInfo": {
-                    "country": "US",
-                    "viewability": "NO_PAGES",
-                    "embeddable": false,
-                    "publicDomain": false,
-                    "textToSpeechPermission": "ALLOWED",
-                    "epub": {
-                        "isAvailable": false
-                    },
-                    "pdf": {
-                        "isAvailable": false
-                    },
-                    "webReaderLink": "http://play.google.com/books/reader?id=PNjzAAAACAAJ&hl=&source=gbs_api",
-                    "accessViewStatus": "NONE",
-                    "quoteSharingAllowed": false
-                }
-            },
-            {
-                "kind": "books#volume",
-                "id": "pnJzAAAACAAJ",
-                "etag": "k8+21p9tjn0",
-                "selfLink": "https://www.googleapis.com/books/v1/volumes/pnJzAAAACAAJ",
-                "volumeInfo": {
-                    "title": "Maritime Fraud",
-                    "authors": [
-                        "Kurt Grönfors"
-                    ],
-                    "publishedDate": "1983",
-                    "industryIdentifiers": [
-                        {
-                            "type": "ISBN_10",
-                            "identifier": "912416190X"
-                        },
-                        {
-                            "type": "ISBN_13",
-                            "identifier": "9789124161903"
-                        }
-                    ],
-                    "readingModes": {
-                        "text": false,
-                        "image": false
-                    },
-                    "pageCount": 151,
-                    "printType": "BOOK",
-                    "categories": [
-                        "Fraud"
-                    ],
-                    "maturityRating": "NOT_MATURE",
-                    "allowAnonLogging": false,
-                    "contentVersion": "preview-1.0.0",
-                    "language": "en",
-                    "previewLink": "http://books.google.com/books?id=pnJzAAAACAAJ&dq=/pnJZAAAAcAAJ&hl=&cd=3&source=gbs_api",
-                    "infoLink": "http://books.google.com/books?id=pnJzAAAACAAJ&dq=/pnJZAAAAcAAJ&hl=&source=gbs_api",
-                    "canonicalVolumeLink": "https://books.google.com/books/about/Maritime_Fraud.html?hl=&id=pnJzAAAACAAJ"
-                },
-                "saleInfo": {
-                    "country": "US",
-                    "saleability": "NOT_FOR_SALE",
-                    "isEbook": false
-                },
-                "accessInfo": {
-                    "country": "US",
-                    "viewability": "NO_PAGES",
-                    "embeddable": false,
-                    "publicDomain": false,
-                    "textToSpeechPermission": "ALLOWED",
-                    "epub": {
-                        "isAvailable": false
-                    },
-                    "pdf": {
-                        "isAvailable": false
-                    },
-                    "webReaderLink": "http://play.google.com/books/reader?id=pnJzAAAACAAJ&hl=&source=gbs_api",
-                    "accessViewStatus": "NONE",
-                    "quoteSharingAllowed": false
-                }
-            },
-            {
-                "kind": "books#volume",
-                "id": "pnJZAAAAcAAJ",
-                "etag": "uIEJzv0oehY",
-                "selfLink": "https://www.googleapis.com/books/v1/volumes/pnJZAAAAcAAJ",
-                "volumeInfo": {
-                    "title": "Album de Azara. Corona cientifica, literaria, artistica y politica que las universidades academias ... cuerpos científicos y patrioticos ... y hombres políticos ... consagran á la buena memoria del ... Señor D. Jose Nicolas de Azara y Perera ... Obra escrita en parte, y dirigida en lo demás por D. B. S. Castellanos de Losada. [With portrait.]",
-                    "authors": [
-                        "Basilio Sebastian CASTELLANOS DE LOSADA"
-                    ],
-                    "publishedDate": "1857",
-                    "industryIdentifiers": [
-                        {
-                            "type": "OTHER",
-                            "identifier": "BL:A0018102859"
-                        }
-                    ],
-                    "readingModes": {
-                        "text": false,
-                        "image": true
-                    },
-                    "pageCount": 770,
-                    "printType": "BOOK",
-                    "maturityRating": "NOT_MATURE",
-                    "allowAnonLogging": false,
-                    "contentVersion": "0.6.6.0.full.1",
-                    "panelizationSummary": {
-                        "containsEpubBubbles": false,
-                        "containsImageBubbles": false
-                    },
-                    "imageLinks": {
-                        "smallThumbnail": "http://books.google.com/books/content?id=pnJZAAAAcAAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
-                        "thumbnail": "http://books.google.com/books/content?id=pnJZAAAAcAAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api"
-                    },
-                    "language": "en",
-                    "previewLink": "http://books.google.com/books?id=pnJZAAAAcAAJ&printsec=frontcover&dq=/pnJZAAAAcAAJ&hl=&cd=4&source=gbs_api",
-                    "infoLink": "https://play.google.com/store/books/details?id=pnJZAAAAcAAJ&source=gbs_api",
-                    "canonicalVolumeLink": "https://play.google.com/store/books/details?id=pnJZAAAAcAAJ"
-                },
-                "saleInfo": {
-                    "country": "US",
-                    "saleability": "FREE",
-                    "isEbook": true,
-                    "buyLink": "https://play.google.com/store/books/details?id=pnJZAAAAcAAJ&rdid=book-pnJZAAAAcAAJ&rdot=1&source=gbs_api"
-                },
-                "accessInfo": {
-                    "country": "US",
-                    "viewability": "ALL_PAGES",
-                    "embeddable": true,
-                    "publicDomain": true,
-                    "textToSpeechPermission": "ALLOWED",
-                    "epub": {
-                        "isAvailable": false,
-                        "downloadLink": "http://books.google.com/books/download/Album_de_Azara_Corona_cientifica_literar.epub?id=pnJZAAAAcAAJ&hl=&output=epub&source=gbs_api"
-                    },
-                    "pdf": {
-                        "isAvailable": false
-                    },
-                    "webReaderLink": "http://play.google.com/books/reader?id=pnJZAAAAcAAJ&hl=&source=gbs_api",
-                    "accessViewStatus": "FULL_PUBLIC_DOMAIN",
-                    "quoteSharingAllowed": false
-                }
-            },
-            {
-                "kind": "books#volume",
-                "id": "pNjzAAAACAAJ",
-                "etag": "qKnupS65EeI",
-                "selfLink": "https://www.googleapis.com/books/v1/volumes/pNjzAAAACAAJ",
-                "volumeInfo": {
-                    "title": "Deutsch-russisch-jakutisches phraseologisches Wörterbuch",
-                    "publishedDate": "2001",
-                    "industryIdentifiers": [
-                        {
-                            "type": "ISBN_10",
-                            "identifier": "5934061836"
-                        },
-                        {
-                            "type": "ISBN_13",
-                            "identifier": "9785934061839"
-                        }
-                    ],
-                    "readingModes": {
-                        "text": false,
-                        "image": false
-                    },
-                    "pageCount": 100,
-                    "printType": "BOOK",
-                    "categories": [
-                        "German language"
-                    ],
-                    "maturityRating": "NOT_MATURE",
-                    "allowAnonLogging": false,
-                    "contentVersion": "preview-1.0.0",
-                    "language": "de",
-                    "previewLink": "http://books.google.com/books?id=pNjzAAAACAAJ&dq=/pnJZAAAAcAAJ&hl=&cd=5&source=gbs_api",
-                    "infoLink": "http://books.google.com/books?id=pNjzAAAACAAJ&dq=/pnJZAAAAcAAJ&hl=&source=gbs_api",
-                    "canonicalVolumeLink": "https://books.google.com/books/about/Deutsch_russisch_jakutisches_phraseologi.html?hl=&id=pNjzAAAACAAJ"
-                },
-                "saleInfo": {
-                    "country": "US",
-                    "saleability": "NOT_FOR_SALE",
-                    "isEbook": false
-                },
-                "accessInfo": {
-                    "country": "US",
-                    "viewability": "NO_PAGES",
-                    "embeddable": false,
-                    "publicDomain": false,
-                    "textToSpeechPermission": "ALLOWED",
-                    "epub": {
-                        "isAvailable": false
-                    },
-                    "pdf": {
-                        "isAvailable": false
-                    },
-                    "webReaderLink": "http://play.google.com/books/reader?id=pNjzAAAACAAJ&hl=&source=gbs_api",
-                    "accessViewStatus": "NONE",
-                    "quoteSharingAllowed": false
-                }
-            },
-            {
-                "kind": "books#volume",
-                "id": "pNJZAAAAcAAJ",
-                "etag": "eFOls5b48Zs",
-                "selfLink": "https://www.googleapis.com/books/v1/volumes/pNJZAAAAcAAJ",
-                "volumeInfo": {
-                    "title": "Jahres-Bericht über das K. K. Josefstädter Ober-Gymnasium",
-                    "subtitle": "über das Schuljahr .... 1872",
-                    "authors": [
-                        "Kaiserlich-Königliches Josefstädter Obergymnasium"
-                    ],
-                    "publishedDate": "1872",
-                    "industryIdentifiers": [
-                        {
-                            "type": "OTHER",
-                            "identifier": "BSB:BSB11041789"
-                        }
-                    ],
-                    "readingModes": {
-                        "text": false,
-                        "image": true
-                    },
-                    "pageCount": 154,
-                    "printType": "BOOK",
-                    "maturityRating": "NOT_MATURE",
-                    "allowAnonLogging": false,
-                    "contentVersion": "0.6.5.0.full.1",
-                    "panelizationSummary": {
-                        "containsEpubBubbles": false,
-                        "containsImageBubbles": false
-                    },
-                    "imageLinks": {
-                        "smallThumbnail": "http://books.google.com/books/content?id=pNJZAAAAcAAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
-                        "thumbnail": "http://books.google.com/books/content?id=pNJZAAAAcAAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api"
-                    },
-                    "language": "de",
-                    "previewLink": "http://books.google.com/books?id=pNJZAAAAcAAJ&printsec=frontcover&dq=/pnJZAAAAcAAJ&hl=&cd=6&source=gbs_api",
-                    "infoLink": "https://play.google.com/store/books/details?id=pNJZAAAAcAAJ&source=gbs_api",
-                    "canonicalVolumeLink": "https://play.google.com/store/books/details?id=pNJZAAAAcAAJ"
-                },
-                "saleInfo": {
-                    "country": "US",
-                    "saleability": "FREE",
-                    "isEbook": true,
-                    "buyLink": "https://play.google.com/store/books/details?id=pNJZAAAAcAAJ&rdid=book-pNJZAAAAcAAJ&rdot=1&source=gbs_api"
-                },
-                "accessInfo": {
-                    "country": "US",
-                    "viewability": "ALL_PAGES",
-                    "embeddable": true,
-                    "publicDomain": true,
-                    "textToSpeechPermission": "ALLOWED",
-                    "epub": {
-                        "isAvailable": false,
-                        "downloadLink": "http://books.google.com/books/download/Jahres_Bericht_%C3%BCber_das_K_K_Josefst%C3%A4dt.epub?id=pNJZAAAAcAAJ&hl=&output=epub&source=gbs_api"
-                    },
-                    "pdf": {
-                        "isAvailable": false
-                    },
-                    "webReaderLink": "http://play.google.com/books/reader?id=pNJZAAAAcAAJ&hl=&source=gbs_api",
-                    "accessViewStatus": "FULL_PUBLIC_DOMAIN",
-                    "quoteSharingAllowed": false
-                }
-            }
-        ]
-    }
-}
