+++ frontend/src/pages/Books.jsx
@@ -1,13 +1,12 @@
+import CategoriesNav from "../components/categories/CategoriesNav";
 import BooksContainer from "../components/book/BooksContainer";
 
-const Library = () => {
+const Books = () => {
   return (
     <div className="col-start-3 col-span-full">
+      <CategoriesNav />
       <BooksContainer />
     </div>
   );
 };
-export default Library;
-
-// the page is intended for testing purposes
-// a margin top 96 px added globally for sections in main css
+export default Books;
